Rename DealDamage to DealDmg, shorter
