+++ src/actions/attacks/slash.py
@@ -1,7 +1,7 @@
 from src.common.enums import Direction
 from src.common.point import Point
 from src.actions.action import Action
-from src.actions.deal_damage import DealDmg
+from src.actions.deal_dmg import DealDmg
 from src.common.serializable import Properties
 from src.objects.object import Object
 from src.objects.npcs.enemy import Enemy
